Move electron to main directory and react app to renderer directory
