+++ src/renderer/components/FileList/Preview/index.jsx
