+++ src/renderer/components/FileList/Preview/index.module.css
