+++ src/renderer/components/MediaControls/ElapsedTimeBar/index.jsx
