+++ src/renderer/components/MediaControls/ElapsedTimeBar/index.module.css
