+++ src/renderer/components/MediaControls/index.jsx
