+++ src/renderer/components/MediaControls/index.module.css
