+++ src/renderer/components/MediaDisplayer/index.jsx
