+++ src/renderer/components/Menu/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import cx from 'clsx'
 import { FormattedMessage, useIntl } from 'react-intl'
-import { version } from '../../../package.json'
+import { version } from '../../../../package.json'
 
 import MenuItem from './MenuItem'
 import styles from './index.module.css'
