+++ src/renderer/translations/es.json
