+++ src/renderer/translations/pt.json
